+++ src/components/shop/ShopDataForm.tsx
@@ -1,12 +1,14 @@
-import { zodResolver } from "@hookform/resolvers/zod";
 import Image from "next/image";
-import { ChangeEvent, useState } from "react";
-import { useForm } from "react-hook-form";
-import * as z from "zod";
+import Link from "next/link";
 
-import { postImages, postShopRegistData, putPresignedURL } from "@/apis/shops";
 import ShopImageCard from "@/components/shop/register/ShopImageCard";
-import { ShopRegisterFormModal } from "@/components/shop/register/ShopRegisterFormModal";
+import {
+  AlertDialogAction,
+  AlertDialogContent,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from "@/components/ui/alert-dialog";
 import { AlertDialog, AlertDialogTrigger } from "@/components/ui/alert-dialog";
 import { Button } from "@/components/ui/button";
 import {
@@ -64,57 +66,26 @@ const ADDRESS = [
   "서울시 강동구",
 ];
 
-const formSchema = z.object({
-  name: z.string(),
-  category: z.string(),
-  address1: z.string(),
-  address2: z.string(),
-  description: z.string(),
-  imageUrl: z.string(),
-  originalHourlyPay: z.string(),
-});
+interface ShopDataFormProps {
+  form: any;
+  onSubmit: (values: any) => void;
+  imgURL: string;
+  handleInputImgFile: (value: any) => void;
+  buttonText: string;
+  modalData: {
+    msg: string;
+    path: string;
+  } | null;
+}
 
-export default function ShopRegisterForm() {
-  const [responseResult, setResponseResult] = useState("");
-  const [imgURL, setImgURL] = useState("");
-  const form = useForm<z.infer<typeof formSchema>>({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      name: "",
-      category: "",
-      address1: "",
-      address2: "",
-      description: "",
-      imageUrl: "https://i.ibb.co/0V2PH9f/default.jpg",
-      originalHourlyPay: "",
-    },
-  });
-  const handleInputImgFile = async (e: ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (file) {
-      const presignedURL = await postImages(token, file.name);
-      await putPresignedURL(presignedURL, file);
-      const nonQueryPresignedURL = presignedURL.split("?")[0];
-      setImgURL(nonQueryPresignedURL);
-      form.setValue("imageUrl", imgURL);
-    }
-  };
-
-  const token = ""; // 임시값
-
-  async function onSubmit(values: z.infer<typeof formSchema>) {
-    try {
-      postShopRegistData(token, values);
-      setResponseResult("200");
-    } catch (e: any) {
-      if (e.response.status === 401) {
-        setResponseResult("401");
-      } else if (e.response.status === 409) {
-        setResponseResult("409");
-      }
-    }
-  }
-
+export default function ShopDataForm({
+  form,
+  onSubmit,
+  imgURL,
+  handleInputImgFile,
+  buttonText,
+  modalData,
+}: ShopDataFormProps) {
   return (
     <>
       <div className="flex justify-between">
@@ -251,9 +222,22 @@ export default function ShopRegisterForm() {
           />
           <AlertDialog>
             <AlertDialogTrigger asChild>
-              <Button type="submit">등록하기</Button>
+              <Button type="submit">{buttonText}</Button>
             </AlertDialogTrigger>
-            <ShopRegisterFormModal responseResult={responseResult} />
+            <AlertDialogContent>
+              {modalData && (
+                <>
+                  <AlertDialogHeader>
+                    <AlertDialogTitle>{modalData.msg}</AlertDialogTitle>
+                  </AlertDialogHeader>
+                  <AlertDialogFooter>
+                    <Link href={modalData.path}>
+                      <AlertDialogAction>확인</AlertDialogAction>
+                    </Link>
+                  </AlertDialogFooter>
+                </>
+              )}
+            </AlertDialogContent>
           </AlertDialog>
         </form>
       </Form>
